[chore] refactor tests
